Add .gitignore entries and increase line-length in pyproject.toml
The commit adds .gitignore entries for "lightning_logs/" and increases the line-length in pyproject.toml from 88 to 90 to match the default behavior of "black" code formatter.

In the dataset folder, the commit introduces a new __init__.py file that exports the PCAMDataset class. Additionally, a new datamodule.py file is added that defines a PyTorch Lightning DataModule for the PCAM dataset.

In the dataset.py file, a new conditional statement is added to apply a default transformation if no transform is specified.

Lastly, an empty __init__.py file is added to the models folder.