feat: Add support for cropping center region of image
This commit adds support for cropping the center 32x32 pixel region of the image in the PCAM dataset. The `crop_center` argument has been added to the `PCAMDataModule` and `PCAMDataset` classes, and a corresponding flag has been added to the command line arguments. The cropped region is extracted in the `PCAMDataset` class by calling `crop` on the image sample with the specified coordinates. In the `SimpleCNN` model, the size of the feature maps in the classifier layer has been adjusted accordingly based on the input size.